avcodec/h264_cabac: use int instead of long for mbb_xy
The mb address fits in int

(cherry picked from commit 592ba6ec106206f97133c9345313010c76361e12)

+++ libavcodec/h264_cabac.c
@@ -1282,7 +1282,7 @@ void ff_h264_init_cabac_states(H264Context *h) {
 }
 
 static int decode_cabac_field_decoding_flag(H264Context *h) {
-    const long mbb_xy = h->mb_xy - 2L*h->mb_stride;
+    const int mbb_xy = h->mb_xy - 2*h->mb_stride;
 
     unsigned long ctx = 0;
